Refactor Table and Search-bar to accept null nodes
Sometimes we want to write code similar to
{booleanValue && <ReactNode/>} on places where we want to render
something conditionally. On false cases, the returned null or false or
undefined value wont render because React knows what to do.

However, in the case of table columns, we dont use the children
<Table.Column> elements as render functions, but as config instead,
to iterate over all columns we use React.Children map or table.props.children.map
and those iterator will list null values as items. This prevents the use
of the {false && <Table.Column />} trick to conditionally render table columns.

This patch filters out nullish values on the places of table and
search-bar components that loops over table children. And adds support
for this style of writing.

Fixes: https://gitlab.com/stoqtech/private/stoq.link/issues/8
Change-Id: I14773783e8af175d3fc387c183be639d33cba748